requirements.txt
mcp-python>=0.1.0
pydantic>=2.0.0
typing-extensions>=4.7.0
asyncio>=3.4.3